refactor: lots of file moving, add README
