--- app/helpers.py
@@ -1,5 +1,0 @@
-"""
-Helper Functions Module
-"""
-
-
