--- app/templates/categories.html
@@ -1,64 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-Replace the title
-{% endblock %}
-{%block main %}
-<div class="grid grid-cols-2 gap-4 items-start justify-items-center mx-auto">
-    <div class="bg-white shadow rounded-lg p-6 mb-8">
-        <h2 class="text-xl font-semibold mb-4">Add New Category</h2>
-        <form action="/categories/add" method="post">
-            <div class="mb-4">
-                <label for="category-name" class="block text-gray-700 text-sm font-bold mb-2">Category
-                    Name:</label>
-                <input type="text" id="category-name" name="category-name"
-                    class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                    placeholder="e.g., Programming Languages" />
-            </div>
-            <button type="submit" class="bg-indigo-800 hover:bg-indigo-700 text-white font-bold py-2 px-4 rounded">
-                Add Category
-            </button>
-        </form>
-    </div>
-    {% if categories %}
-    <div class="bg-white shadow rounded-lg p-6 mb-8">
-        <h2 class="text-xl font-semibold mb-4">Catories Overview</h2>
-        <table class="min-w-full divide-y divide-gray-200">
-            <thead>
-                <tr>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        ID
-                    </th>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Category
-                    </th>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Actions
-                    </th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for category in categories %}
-                <tr class="bg-white">
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                        {{ category.id }}
-                    </td>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                        {{ category.name}}
-                    </td>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                        <form action="/categories/delete/{{ category.id }}" method="post">
-                            <button type="submit"
-                                class="bg-indigo-800 hover:bg-indigo-700 text-white font-bold py-2 px-4 rounded">
-                                Delete
-                            </button>
-                        </form>
-                    </td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-    {% endif %}
-</div>
-
-{% endblock %}
