--- app/templates/skills.html
@@ -1,70 +1,0 @@
-{% extends "base.html" %}
-{% block title %} Skills View {% endblock %}
-{% block main %}
-<div class="grid grid-cols-1 gap-4 items-start justify-items-center mx-auto">
-    <div class="bg-white shadow rounded-lg p-6 mb-8 w-full">
-        <h2 class="text-xl font-semibold mb-4">Add New Skill</h2>
-        <form action="/skills/add" method="post">
-            <div class="mb-4">
-                <label for="skill-name" class="block text-gray-700 text-sm font-bold mb-2">Skill Name:</label>
-                <input autofocus type="text" id="skill-name" name="skill-name"
-                    class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                    placeholder="e.g., Python" />
-            </div>
-            <button type="submit" class="bg-indigo-800 hover:bg-indigo-700 text-white font-bold py-2 px-4 rounded">
-                Add Skill
-            </button>
-        </form>
-    </div>
-    
-    <div class="bg-white shadow rounded-lg p-6 mb-8 w-full">
-        <h2 class="text-xl font-semibold mb-4">Skills Overview</h2>
-        {% if skills %}
-        <table class="min-w-full divide-y divide-gray-200">
-            <thead>
-                <tr>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Skill</th>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Actions</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for skill in skills %}
-                <tr class="bg-white">
-                    <td class="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{{ skill.name }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                        <!-- Add Parent -->
-                        <form action="/skills/{{ skill.id }}/add-parent" method="post" class="inline">
-                            <select name="parent-id" required
-                                class="shadow appearance-none border rounded py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                                title="addparent">
-                                <option value="" disabled selected>Choose Parent</option>
-                                {% for potential_parent in skills %}
-                                {% if potential_parent.id != skill.id %}
-                                <option value="{{ potential_parent.id }}">{{ potential_parent.name }}</option>
-                                {% endif %}
-                                {% endfor %}
-                            </select>
-                            <button type="submit"
-                                class="bg-indigo-800 hover:bg-indigo-700 text-white font-bold py-2 px-4 rounded">
-                                Add Parent
-                            </button>
-                        </form>
-                        <!-- Delete Skill -->
-                        <form action="/skills/delete/{{ skill.id }}" method="post" class="inline">
-                            <button type="submit"
-                                class="bg-red-600 hover:bg-red-500 text-white font-bold py-2 px-4 rounded">
-                                Delete
-                            </button>
-                        </form>
-                    </td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-        {% else %}
-        <p class="text-gray-700">No skills available. Add a new skill above!</p>
-    {% endif %}
-</div>
-</div>
-
-{% endblock %}
