+++ frontend/src/form_components/buttons/index.js
@@ -1,0 +1,6 @@
+import TextWithIcon from "./TextWithIcon";
+import TextWithNA from "./TextWithNA";
+import TextWithPlus from "./TextWithPlus";
+import TextWithX from "./TextWithX";
+
+export { TextWithIcon, TextWithNA, TextWithPlus, TextWithX };
