+++ frontend/src/form_components/skill_form/SkillParentList.jsx
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import TextWithPlus from "./TextWithPlus";
+import TextWithPlus from "../buttons/TextWithPlus";
 import PropTypes from 'prop-types';
 
 const SkillParentList = ({ Component = TextWithPlus, onClick, skillList }) => {
