+++ requirements.txt
@@ -2,3 +2,4 @@ flask
 flask-cors
 flask-migrate
 flask-sqlalchemy
+requests
